requirements.txt
django==2.2.7
psycopg2==2.8.4
djangorestframework==3.10.3
whitenoise==4.1.4